requirements.txt
cachetools==5.0.0
google-auth==2.6.2
google-auth-oauthlib==0.5.1
gspread==5.2.0
numpy==1.22.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
